Relocate manager functions
